fix: refactored baseDeployRetrieve.ts
